+++ app/core/auth.c
@@ -8,8 +8,7 @@
 #include "storage/keys.h"
 #include "ui/ui.h"
 #include "ur/ur.h"
-#include "ur/auth_types.h"
-#include "ur/auth_encode.h"
+#include "ur/ur_encode.h"
 
 #define AUTH_CHALLENGE_LEN 32
 #define AUTH_SIG_LEN 64
@@ -24,11 +23,11 @@ void device_auth_run() {
 
   if ((auth.dev_auth_step.dev_auth_step_type_choice != dev_auth_step_type_dev_auth_init_m_c) ||
       !auth.dev_auth_challenge_present ||
-      auth.dev_auth_device_id_present ||
+      auth.dev_auth_dev_id_present ||
       auth.dev_auth_first_auth_present ||
       auth.dev_auth_auth_time_present ||
       auth.dev_auth_auth_count_present ||
-      auth.dev_auth_signature_present) {
+      auth.dev_auth_auth_sig_present) {
     ui_info(LSTR(DEV_AUTH_INVALID_CHALLENGE), 1);
     return;
   }
@@ -53,14 +52,14 @@ void device_auth_run() {
   // Response
   random_buffer((uint8_t*) auth.dev_auth_challenge.dev_auth_challenge.value, AUTH_CHALLENGE_LEN);
 
-  auth.dev_auth_device_id_present = 1;
-  auth.dev_auth_device_id.dev_auth_device_id.value = uid;
-  auth.dev_auth_device_id.dev_auth_device_id.len = HAL_DEVICE_UID_LEN;
+  auth.dev_auth_dev_id_present = 1;
+  auth.dev_auth_dev_id.dev_auth_dev_id.value = uid;
+  auth.dev_auth_dev_id.dev_auth_dev_id.len = HAL_DEVICE_UID_LEN;
 
   auth.dev_auth_step.dev_auth_step_type_choice = dev_auth_step_type_dev_auth_device_m_c;
-  auth.dev_auth_signature_present = 1;
-  auth.dev_auth_signature.dev_auth_signature.value = g_core.data.sig.plain_sig;
-  auth.dev_auth_signature.dev_auth_signature.len = AUTH_SIG_LEN;
+  auth.dev_auth_auth_sig_present = 1;
+  auth.dev_auth_auth_sig.dev_auth_auth_sig.value = g_core.data.sig.plain_sig;
+  auth.dev_auth_auth_sig.dev_auth_auth_sig.len = AUTH_SIG_LEN;
 
   cbor_encode_dev_auth(g_core.data.sig.cbor_sig, CBOR_SIG_MAX_LEN, &auth, &g_core.data.sig.cbor_len);
 
@@ -77,11 +76,11 @@ void device_auth_run() {
 
   if ((auth.dev_auth_step.dev_auth_step_type_choice != dev_auth_step_type_dev_auth_server_m_c) ||
       auth.dev_auth_challenge_present ||
-      auth.dev_auth_device_id_present ||
+      auth.dev_auth_dev_id_present ||
       !auth.dev_auth_first_auth_present ||
       !auth.dev_auth_auth_time_present ||
       !auth.dev_auth_auth_count_present ||
-      !auth.dev_auth_signature_present
+      !auth.dev_auth_auth_sig_present
       ) {
     ui_info(LSTR(DEV_AUTH_INVALID_CHALLENGE), 1);
     return;
@@ -94,7 +93,7 @@ void device_auth_run() {
 
   const uint8_t* key;
   key_read_public(DEV_AUTH_SERVER_KEY, &key);
-  if (!ecdsa_verify(&secp256k1, key, auth.dev_auth_signature.dev_auth_signature.value, digest)) {
+  if (!ecdsa_verify(&secp256k1, key, auth.dev_auth_auth_sig.dev_auth_auth_sig.value, digest)) {
     ui_device_auth(auth.dev_auth_first_auth.dev_auth_first_auth, auth.dev_auth_auth_time.dev_auth_auth_time, auth.dev_auth_auth_count.dev_auth_auth_count);
   } else {
     ui_info(LSTR(DEV_AUTH_INVALID_AUTH), 1);
