+++ app/core/core.h
@@ -12,7 +12,7 @@
 #include "keycard/keycard.h"
 #include "iso7816/smartcard.h"
 #include "ui/ui.h"
-#include "ur/eip4527_types.h"
+#include "ur/ur_types.h"
 
 #define BIP44_MAX_PATH_LEN 40
