--- app/ur/btc_types.h
@@ -1,34 +1,0 @@
-/*
- * Generated using zcbor version 0.8.1
- * https://github.com/NordicSemiconductor/zcbor
- * Generated with a --default-max-qty of 3
- */
-
-#ifndef BTC_TYPES_H__
-#define BTC_TYPES_H__
-
-#include <stdint.h>
-#include <stdbool.h>
-#include <stddef.h>
-
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/** Which value for --default-max-qty this file was created with.
- *
- *  The define is used in the other generated file to do a build-time
- *  compatibility check.
- *
- *  See `zcbor --help` for more information about --default-max-qty
- */
-#define DEFAULT_MAX_QTY 3
-
-
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif /* BTC_TYPES_H__ */
