+++ app/ur/ur.h
@@ -10,8 +10,8 @@
 typedef enum {
   ETH_SIGNATURE = 0,
   CRYPTO_OUTPUT = 1,
-  CRYPTO_KEYPATH = 2,
-  ETH_SIGN_REQUEST = 3,
+  CRYPTO_ACCOUNT = 2,
+  ETH_SIGN_REQUEST = 4,
   BYTES = 6,
   FS_DATA = 8,
   DEV_AUTH = 9,
