--- app/ur/ur_part_decode.c
@@ -1,56 +1,0 @@
-/*
- * Generated using zcbor version 0.8.1
- * https://github.com/NordicSemiconductor/zcbor
- * Generated with a --default-max-qty of 3
- */
-
-#include <stdint.h>
-#include <stdbool.h>
-#include <stddef.h>
-#include <string.h>
-#include "zcbor_decode.h"
-#include "ur_part_decode.h"
-#include "zcbor_print.h"
-
-#if DEFAULT_MAX_QTY != 3
-#error "The type file was generated with a different default_max_qty than this file"
-#endif
-
-static bool decode_ur_part(zcbor_state_t *state, struct ur_part *result);
-
-
-static bool decode_ur_part(
-		zcbor_state_t *state, struct ur_part *result)
-{
-	zcbor_log("%s\r\n", __func__);
-
-	bool tmp_result = (((zcbor_list_start_decode(state) && ((((zcbor_uint32_decode(state, (&(*result).ur_part_seqNum)))
-	&& ((((((*result).ur_part_seqNum <= UINT32_MAX)) || (zcbor_error(state, ZCBOR_ERR_WRONG_RANGE), false))) || (zcbor_error(state, ZCBOR_ERR_WRONG_RANGE), false)))
-	&& ((zcbor_uint32_decode(state, (&(*result).ur_part_seqLen))))
-	&& ((zcbor_uint32_decode(state, (&(*result).ur_part_messageLen))))
-	&& ((zcbor_uint32_decode(state, (&(*result).ur_part_checksum)))
-	&& ((((((*result).ur_part_checksum <= UINT32_MAX)) || (zcbor_error(state, ZCBOR_ERR_WRONG_RANGE), false))) || (zcbor_error(state, ZCBOR_ERR_WRONG_RANGE), false)))
-	&& ((zcbor_bstr_decode(state, (&(*result).ur_part_data))))) || (zcbor_list_map_end_force_decode(state), false)) && zcbor_list_end_decode(state))));
-
-	if (!tmp_result) {
-		zcbor_trace_file(state);
-		zcbor_log("%s error: %s\r\n", __func__, zcbor_error_str(zcbor_peek_error(state)));
-	} else {
-		zcbor_log("%s success\r\n", __func__);
-	}
-
-	return tmp_result;
-}
-
-
-
-int cbor_decode_ur_part(
-		const uint8_t *payload, size_t payload_len,
-		struct ur_part *result,
-		size_t *payload_len_out)
-{
-	zcbor_state_t states[3];
-
-	return zcbor_entry_function(payload, payload_len, (void *)result, payload_len_out, states,
-		(zcbor_decoder_t *)decode_ur_part, sizeof(states) / sizeof(zcbor_state_t), 1);
-}
