--- cddl/btc.cddl
@@ -1,1 +1,0 @@
-psbt = bstr
